fix: sh

+++ sh/qa.sh
@@ -9,10 +9,20 @@ ls
 pnpm -v
 pnpm install
 pnpm run build
-ls -a .next
 
-rm -rf node_modules
+if [ $? -eq 0 ]
+then
+  ls -a .next
 
-rm -rf build.tar.gz
-tar --exclude=build.tar.gz  --exclude=.git --warning=no-file-changed -zcvf build.tar.gz ./
-echo 'success'
+  rm -rf node_modules
+
+  rm -rf build.tar.gz
+  tar --exclude=build.tar.gz  --exclude=.git --warning=no-file-changed -zcvf build.tar.gz ./
+  echo 'success'
+  exit 0
+else
+  echo "Build Failed" >&2
+  exit 1
+fi
+
+
